Add profiling options to command line launching
Adds python profiling to mantid using the cProfile module. Lauching
workbench with a --profile modifer and a file path makes workbench
output a file to the given path with that can be analysed with snakeviz.

RE #28499
