+++ buildconfig/CMake/Packaging/launch_workbench.pyw
@@ -1,4 +1,4 @@
 # Small startup script that replaces the compiled executable from setuptools
 # This script should be called with Mantid's Python from the launch_workbench run script
-import workbench.app.main
-workbench.app.main.main()
+import workbench.app.start
+workbench.app.start.main()
